% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_data_export.R
\name{add_summary_sheet}
\alias{add_summary_sheet}
\title{Add summary sheet to Excel workbook}
\usage{
add_summary_sheet(wb, data, main_sheet_name)
}
\arguments{
\item{wb}{Workbook object}

\item{data}{Data frame}

\item{main_sheet_name}{Character string for main worksheet name}
}
\value{
Invisible NULL
}
\description{
Internal function to add a summary sheet with data statistics.
}
