% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_development_tools.R
\name{generate_coding_standards}
\alias{generate_coding_standards}
\title{Generate Coding Standards Documentation}
\usage{
generate_coding_standards(output_dir, package_name, overwrite)
}
\description{
Generate Coding Standards Documentation
}
\keyword{internal}
