% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_development_tools.R
\name{generate_documentation_standards}
\alias{generate_documentation_standards}
\title{Generate Documentation Standards Documentation}
\usage{
generate_documentation_standards(output_dir, package_name, overwrite)
}
\description{
Generate Documentation Standards Documentation
}
\keyword{internal}
