% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_development_tools.R
\name{generate_git_workflow}
\alias{generate_git_workflow}
\title{Generate Git Workflow Documentation}
\usage{
generate_git_workflow(output_dir, package_name, overwrite)
}
\description{
Generate Git Workflow Documentation
}
\keyword{internal}
