% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_development_tools.R
\name{generate_project_overview}
\alias{generate_project_overview}
\title{Generate Project Overview Documentation}
\usage{
generate_project_overview(output_dir, package_name, overwrite)
}
\description{
Generate Project Overview Documentation
}
\keyword{internal}
