% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_development_tools.R
\name{generate_project_templates}
\alias{generate_project_templates}
\title{Generate Project Templates}
\usage{
generate_project_templates(output_dir, package_name, overwrite)
}
\description{
Generate Project Templates
}
\keyword{internal}
