% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_development_tools.R
\name{generate_release_checklist}
\alias{generate_release_checklist}
\title{Generate Release Checklist Documentation}
\usage{
generate_release_checklist(output_dir, package_name, overwrite)
}
\description{
Generate Release Checklist Documentation
}
\keyword{internal}
