% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_development_tools.R
\name{generate_templates}
\alias{generate_templates}
\title{Generate Templates}
\usage{
generate_templates(output_dir, package_name, overwrite)
}
\description{
Generate Templates
}
\keyword{internal}
