% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_development_tools.R
\name{generate_testing_standards}
\alias{generate_testing_standards}
\title{Generate Testing Standards Documentation}
\usage{
generate_testing_standards(output_dir, package_name, overwrite)
}
\description{
Generate Testing Standards Documentation
}
\keyword{internal}
