% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_package_management.R
\name{try_install_bioc}
\alias{try_install_bioc}
\title{Try to install a package from Bioconductor}
\usage{
try_install_bioc(pkg, dependencies, quiet)
}
\description{
Try to install a package from Bioconductor
}
\keyword{internal}
