% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_package_management.R
\name{try_install_cran}
\alias{try_install_cran}
\title{Try to install a package from CRAN}
\usage{
try_install_cran(pkg, repos, dependencies, quiet)
}
\description{
Try to install a package from CRAN
}
\keyword{internal}
